Update preprocessImage.m for GALFIT use
I needed to pull Sparcfire's galaxy crop for use in cropping the original FITS image (to maintain image integrity just in case) so I inserted this code. As noted, it shouldn't affect anything else, cropRem is the only variable not cleared and it's overwritten on line 625 regardless. I meant to make the pull a long time ago but got hung up on some github things and forgot about it.

The only issue I've had with this code is that there's an extra '_crop_coord.txt' file generated in the directory where sparcfire is run from. I'm not sure why it occurs but it can be removed once the run is complete I have no need for it.