Adding useful CSV things
Uploading a separate folder which contains a csv holding information necessary for the generation of PSFs given the appropriate psfield files from SDSS. star_dl.csv only contains information for 900 galaxies but the SQL query used to generate it on SDSS' website is included as a plain text file as an example. 

Two scripts are also included which extract RA and DEC info from the FITS file headers and another which is used to download PSField files. They are not necessary to use but may be useful in the future.