Updating preprocessImage for GALFIT
I initially added code to output a new file with the cropping coordinates but wanted to cut down on output - I (much later) realized Sparcfire was already outputting a crop radius which I could use in conjunction with the measure of the center so I redacted that change. 

The crop radius is being output into ...elps-fit-params but it would be much more efficient file I/O-wise if it's in with the galaxy params so I also (I think correctly) added cropRad to what will eventually be the galaxy params csv.