Update writeGxyParamsToCsv.m
Added cropRad as a name for the output CSV so that the crop radius is output in the csv instead of in a separate file (elps-fits-params.txt). This will reduce file I/O for my galfit algorithm and may also reduce the need for elps-fits... Someone will have to check on that but now everything I need is in the csvs.