Updating...
Updated getGalfitFitQuality.m to cd into a tmp directory before running GALFIT so that it dumps all the unnecessary files in there. Furthermore, now only one GALFIT file will be retained in the final directory (at my own behest) just in case.

I also updated the position of cropRad in writeGxyParamsToCsv.m so it's easier to find later and once we update the reg tests, everything should be good to go.