Update GENERATEFITQUALITY
Since getenv takes the system variable as a string, the if statement was merely evaluating if GENERATEFITQUALITY existed. Converting to a number solves this issue - I also checked if there was a problem when GENERATEFITQUALITY wasn't specified but there isn't, MATLAB simply stores an empty char array which also evaluates to false.