Bug fix and minor changes
reProject -- Unfortunately when image standardization is turned off this script reports an error and no csv is generated. Since I haven't hunted down the bug in the calling script, this will have to do and should do as far as my limited testing is concerned. Also provided some documentation while I was at it.

writeGxyParams... -- I tried modifying this file when trying to fix the image standardization problem to see if anything would come out and realized those changes would probably be a good idea anyway so I'm updating this with them. May have to update reg tests... again. At least for image guiding. Checking now but committing before I forget. It may not be worth it to avoid that hassle