Updates
* control_script -- fixed a couple small things
* notebook_feedme/sparc_to_galfit... -- added an ascii starmask generator since GALFIT dies if the starmask is just the image (output if nothing to mask). This ensures we avoid this scenario by just including the coordinates of the masked pixels. Also fixed defaults to be ints/floats instead of strings to avoid an error if sparcfire input can't be read. + some other quality of life fixes
* remove_stars_with_sextractor -- Converted to python3. Fixed deprecation issue by using imageio and generic_filter. Also quits if usage is incorrect.

General notes: GALFIT is acting quite strange on Ubuntu... I'm getting a segfault when it tries to output the final images. This is something I'm actively trying to fix... also, behavior seems to be a bit different, I'm having trouble replicating some results I got over the summer. I'm not sure why this is either.  Now at least everything is in ship shape!