Created and populated ResultsAnalysis dir.
Essentially split everything in the results analysis notebook into separate functions to make debugging and whatnot easier. There may be some additional bug fixes to the code since I last committed but those should be minimal. I tried grouping functions in files where it felt appropriate so some of the files in the top directory of the ResultsAnalysis module are longer to keep everything logically consistent. Also updated the notebook and confirmed that importing these works.