Added code for increasing font size.
There are so many places/ways to increase font in plotly that this is surely a mess. All the plots now have an increased font size from the default, which can be controlled via kwargs. I have also selectively capitalized axes and title labels. It would be useful to make that another kwarg but in due time.