Added star masking to residual for png output.
This was a highly requested feature that solves issues when converting the residual to PNG with bright sources in the image. There is a TODO in FitsHandlers which suggests a better way to approach this but for now, the bandage works. Also updated generate_png.py to coincide with these additions.